# openapi/sixtysix.yaml
openapi: 3.0.3
info:
  title: go-sixtysix API
  version: 0.1.0
  contact:
    name: Rumen Damyanov
    email: contact@rumenx.com
servers:
  - url: http://localhost:8080
paths:
...
  /sessions:
    get:
      summary: List sessions
      parameters:
        - in: query
          name: game
          schema:
            type: string
        - in: query
          name: offset
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
    post:
      summary: Create a session
      parameters:
        - in: query
          name: game
          required: true
          schema:
            type: string
        - in: query
          name: seed
          schema:
            type: integer
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
...
components:
  schemas:
    Session:
      type: object
      properties:
        id:
          type: string
        gameName:
          type: string
        state:
          type: object
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...